+++ .gitignore
@@ -70,3 +70,4 @@ plugin/pkg/vmturbo/metadata/config_nyc.json
 plugin/pkg/vmturbo/probe/helper/testing_flag.json
 vmt_testing/vmt_testing_config.json
 
+vmt_testing/kubeturbo/kubeturbo
