+++ plugin/cmd/kube-vmturbo/app/kubeturbo_builder.go
@@ -19,20 +19,31 @@ package app
 
 import (
 	"fmt"
+	"io/ioutil"
 	"net"
+	"net/http"
+	"net/http/pprof"
 	"os"
+	"strconv"
 
+	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/latest"
 	"k8s.io/kubernetes/pkg/api/meta"
 	"k8s.io/kubernetes/pkg/client"
 	"k8s.io/kubernetes/pkg/client/clientcmd"
 	clientcmdapi "k8s.io/kubernetes/pkg/client/clientcmd/api"
+	"k8s.io/kubernetes/pkg/client/record"
+	"k8s.io/kubernetes/pkg/healthz"
 	"k8s.io/kubernetes/pkg/master"
-	// "k8s.io/kubernetes/pkg/storage"
-	"k8s.io/kubernetes/pkg/tools"
-	// "k8s.io/kubernetes/pkg/healthz"
 	"k8s.io/kubernetes/pkg/master/ports"
-	// "k8s.io/kubernetes/pkg/util"
+	"k8s.io/kubernetes/pkg/tools"
+	"k8s.io/kubernetes/pkg/util"
+	schedulerbuilder "k8s.io/kubernetes/plugin/cmd/kube-scheduler/app"
+	"k8s.io/kubernetes/plugin/pkg/scheduler"
+	_ "k8s.io/kubernetes/plugin/pkg/scheduler/algorithmprovider"
+	schedulerapi "k8s.io/kubernetes/plugin/pkg/scheduler/api"
+	latestschedulerapi "k8s.io/kubernetes/plugin/pkg/scheduler/api/latest"
+	"k8s.io/kubernetes/plugin/pkg/scheduler/factory"
 
 	"k8s.io/kubernetes/plugin/pkg/vmturbo"
 	"k8s.io/kubernetes/plugin/pkg/vmturbo/conversion"
@@ -42,7 +53,12 @@ import (
 	etcdhelper "k8s.io/kubernetes/plugin/pkg/vmturbo/storage/etcd"
 
 	"github.com/golang/glog"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/spf13/pflag"
+)
+
+var (
+	defaultSchedulerInstance *scheduler.Scheduler
 )
 
 // VMTServer has all the context and params needed to run a Scheduler
@@ -85,12 +101,11 @@ func (s *VMTServer) AddFlags(fs *pflag.FlagSet) {
 
 // Run runs the specified VMTServer.  This should never exit.
 func (s *VMTServer) Run(_ []string) error {
-	fmt.Println(".......... Start run vmt server ..........")
 	if s.Kubeconfig == "" && s.Master == "" {
 		glog.Warningf("Neither --kubeconfig nor --master was specified.  Using default API client.  This might not work.")
 	}
 
-	glog.Infof("Master is %s", s.Master)
+	glog.V(3).Infof("Master is %s", s.Master)
 
 	if s.MetaConfigPath == "" {
 		glog.Fatalf("The path to the VMT config file is not provided.Exiting...")
@@ -163,6 +178,8 @@ func (s *VMTServer) Run(_ []string) error {
 
 	vmtService := vmturbo.NewVMTurboService(vmtConfig)
 
+	vmtService.TurboScheduler.SetDefaultScheduler(defaultSchedulerInstance)
+
 	vmtService.Run()
 
 	select {}
@@ -179,4 +196,103 @@ func newEtcd(client tools.EtcdClient, interfacesFunc meta.VersionInterfacesFunc,
 	simpleCodec.AddKnownTypes(&registry.VMTEvent{})
 	simpleCodec.AddKnownTypes(&registry.VMTEventList{})
 	return etcdhelper.NewEtcdStorage(client, simpleCodec, pathPrefix), nil
+}
+
+type DefaultK8sSchedulerServer struct {
+	*schedulerbuilder.SchedulerServer
+}
+
+func NewDefaultK8sSchedulerServer(vmtServer *VMTServer) *DefaultK8sSchedulerServer {
+	s := schedulerbuilder.SchedulerServer{
+		Port:              ports.VMTPort,
+		Address:           net.ParseIP("127.0.0.1"),
+		AlgorithmProvider: factory.DefaultProvider,
+	}
+	return &DefaultK8sSchedulerServer{&s}
+}
+
+// The same method defined in scheduler server. Create config file for the default kubernetes scheduler.
+func (s DefaultK8sSchedulerServer) createConfig(configFactory *factory.ConfigFactory) (*scheduler.Config, error) {
+	var policy schedulerapi.Policy
+	var configData []byte
+
+	if _, err := os.Stat(s.PolicyConfigFile); err == nil {
+		configData, err = ioutil.ReadFile(s.PolicyConfigFile)
+		if err != nil {
+			return nil, fmt.Errorf("Unable to read policy config: %v", err)
+		}
+		err = latestschedulerapi.Codec.DecodeInto(configData, &policy)
+		if err != nil {
+			return nil, fmt.Errorf("Invalid configuration: %v", err)
+		}
+
+		return configFactory.CreateFromConfig(policy)
+	}
+
+	// if the config file isn't provided, use the specified (or default) provider
+	// check of algorithm provider is registered and fail fast
+	_, err := factory.GetAlgorithmProvider(s.AlgorithmProvider)
+	if err != nil {
+		return nil, err
+	}
+
+	return configFactory.CreateFromProvider(s.AlgorithmProvider)
+}
+
+// Run runs the specified SchedulerServer.  This should never exit. Same code from schduler server.
+func (s *DefaultK8sSchedulerServer) RunDefaultK8sScheduler(_ []string) error {
+	if s.Kubeconfig == "" && s.Master == "" {
+		glog.Warningf("Neither --kubeconfig nor --master was specified.  Using default API client.  This might not work.")
+	}
+
+	// This creates a client, first loading any specified kubeconfig
+	// file, and then overriding the Master flag, if non-empty.
+	kubeconfig, err := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
+		&clientcmd.ClientConfigLoadingRules{ExplicitPath: s.Kubeconfig},
+		&clientcmd.ConfigOverrides{ClusterInfo: clientcmdapi.Cluster{Server: s.Master}}).ClientConfig()
+	if err != nil {
+		return err
+	}
+	kubeconfig.QPS = 20.0
+	kubeconfig.Burst = 30
+
+	kubeClient, err := client.New(kubeconfig)
+	if err != nil {
+		glog.Fatalf("Invalid API configuration: %v", err)
+	}
+
+	go func() {
+		mux := http.NewServeMux()
+		healthz.InstallHandler(mux)
+		if s.EnableProfiling {
+			mux.HandleFunc("/debug/pprof/", pprof.Index)
+			mux.HandleFunc("/debug/pprof/profile", pprof.Profile)
+			mux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
+		}
+		mux.Handle("/metrics", prometheus.Handler())
+
+		server := &http.Server{
+			Addr:    net.JoinHostPort(s.Address.String(), strconv.Itoa(s.Port)),
+			Handler: mux,
+		}
+		glog.Fatal(server.ListenAndServe())
+	}()
+
+	configFactory := factory.NewConfigFactory(kubeClient, util.NewTokenBucketRateLimiter(s.BindPodsQPS, s.BindPodsBurst))
+	config, err := s.createConfig(configFactory)
+	if err != nil {
+		glog.Fatalf("Failed to create scheduler configuration: %v", err)
+	}
+
+	eventBroadcaster := record.NewBroadcaster()
+	config.Recorder = eventBroadcaster.NewRecorder(api.EventSource{Component: "scheduler"})
+	eventBroadcaster.StartLogging(glog.Infof)
+	eventBroadcaster.StartRecordingToSink(kubeClient.Events(""))
+
+	sched := scheduler.New(config)
+	// vmturbo.SetSchedulerInstance(sched)
+
+	defaultSchedulerInstance = sched
+
+	select {}
 }
