+++ plugin/pkg/scheduler/scheduler.go
@@ -188,58 +188,9 @@ func (s *Scheduler) scheduleOne() {
 	})
 }
 
-// The VMTSchedule get destination node from vmt service and bind the corresponding pod to the node.
-// TODO, whether to keep it here or move this to vmt service.`
-// TODO, should make the code cleaner.
-func (s *Scheduler) VMTSchedule(pod *api.Pod, dest string) {
-	if s.config.BindPodsRateLimiter != nil {
-		s.config.BindPodsRateLimiter.Accept()
-	}
-
-	start := time.Now()
-	defer func() {
-		metrics.E2eSchedulingLatency.Observe(metrics.SinceInMicroseconds(start))
-	}()
-	// dest, err := s.config.Algorithm.Schedule(pod, s.config.MinionLister)
-	metrics.SchedulingAlgorithmLatency.Observe(metrics.SinceInMicroseconds(start))
-	// if err != nil {
-	// 	glog.V(1).Infof("Failed to schedule: %+v", pod)
-	// 	s.config.Recorder.Eventf(pod, "FailedScheduling", "%v", err)
-	// 	s.config.Error(pod, err)
-	// 	return
-	// }
-	b := &api.Binding{
-		ObjectMeta: api.ObjectMeta{Namespace: pod.Namespace, Name: pod.Name},
-		Target: api.ObjectReference{
-			Kind: "Node",
-			Name: dest,
-		},
-	}
-
-	// We want to add the pod to the model iff the bind succeeds, but we don't want to race
-	// with any deletions, which happen asynchronously.
-	s.config.Modeler.LockedAction(func() {
-		bindingStart := time.Now()
-		err := s.config.Binder.Bind(b)
-		metrics.BindingLatency.Observe(metrics.SinceInMicroseconds(bindingStart))
-		if err != nil {
-			glog.V(1).Infof("Failed to bind pod: %+v", err)
-			s.config.Recorder.Eventf(pod, "FailedScheduling", "Binding rejected: %v", err)
-			s.config.Error(pod, err)
-			return
-		}
-		s.config.Recorder.Eventf(pod, "Scheduled", "Successfully assigned %v to %v", pod.Name, dest)
-		// tell the model to assume that this binding took effect.
-		assumed := *pod
-		assumed.Spec.NodeName = dest
-		s.config.Modeler.AssumePod(&assumed)
-
-	})
-}
-
 // TODO. This should be removed when the vmt reservation api works.
 // Call the built in schduling algorithm to schedule a pod
-func (s *Scheduler) VMTScheduleHelper(pod *api.Pod) (string, error) {
+func (s *Scheduler) FindDestination(pod *api.Pod) (string, error) {
 	dest, err := s.config.Algorithm.Schedule(pod, s.config.MinionLister)
 	if err != nil {
 		glog.Errorf("Error Scheduling pod %+v", pod)
